feat: add notifications, presence, and blacklist features; update WebSocket and pet/user facades
- Implement WebSocket event listener and configuration updates
- Add notifications controller, DTOs, and service logic
- Introduce blacklisted user model and repository
- Add presence indication DTOs and service handling
- Update pet and user facades with new actions
- Refactor chat and admin controllers for event handling
- Delete old migration for image URL adjustments
- Update database migration for notifications and blacklist

+++ src/main/java/com/petbook/petbook_backend/service/PetService.java
@@ -10,14 +10,18 @@ import com.petbook.petbook_backend.dto.response.PetInfoPublicResponse;
 import com.petbook.petbook_backend.exceptions.rest.PetListingNotFoundException;
 import com.petbook.petbook_backend.exceptions.rest.UnauthorizedUserException;
 import com.petbook.petbook_backend.exceptions.rest.UserNotFoundException;
+import com.petbook.petbook_backend.models.Gender;
 import com.petbook.petbook_backend.models.ImageUrl;
+import com.petbook.petbook_backend.models.NotificationType;
 import com.petbook.petbook_backend.models.Pet;
 import com.petbook.petbook_backend.models.User;
 import com.petbook.petbook_backend.repository.PetImageUrlsRepository;
 import com.petbook.petbook_backend.repository.PetRepository;
 import com.petbook.petbook_backend.repository.UserRepository;
+import com.petbook.petbook_backend.service.events.NotificationEvent;
 import jakarta.validation.constraints.NotNull;
 import lombok.RequiredArgsConstructor;
+import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.data.domain.Example;
 import org.springframework.data.domain.ExampleMatcher;
 import org.springframework.data.domain.Page;
@@ -28,6 +32,7 @@ import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -42,6 +47,7 @@ public class PetService {
     private final PetRepository petRepository;
     private final PetImageUrlsRepository petImageUrlsRepository;
     private final CloudinaryService cloudinaryService;
+    private final ApplicationEventPublisher applicationEventPublisher;
 
 
 //TODO: Change general pet listings query api to only read one imageUrl from the db
@@ -53,7 +59,6 @@ public class PetService {
         String username = SecurityContextHolder.getContext().getAuthentication().getName();
         User user = userRepository.findByEmail(username)
                 .orElseThrow(() -> new UserNotFoundException("User not found"));
-        String email = user.getEmail();
 
         Pet pet = petRepository.findById(petId)
                 .orElseThrow(() -> new PetListingNotFoundException("Listing not found:" + petId));
@@ -68,7 +73,15 @@ public class PetService {
     public PetInfoPublicResponse getPetById(Long id) {
         Pet pet = petRepository.findByIdAndApproved(id, true).orElseThrow(() ->
                 new PetListingNotFoundException("No Pet listing found for given id"));
-        String ownerEmail = (pet.getOwner() != null) ? pet.getOwner().getEmail() : null;
+        String ownerEmail;
+        Long ownerId;
+        if (pet.getOwner() != null) {
+            ownerEmail = pet.getOwner().getEmail();
+            ownerId = pet.getOwner().getId();
+        } else {
+            ownerEmail = null;
+            ownerId = null;
+        }
         List<ImageUrl> imageUrls = pet.getImages();
         return PetInfoPublicResponse.builder()
                 .id(pet.getId())
@@ -76,9 +89,12 @@ public class PetService {
                 .type(pet.getType())
                 .breed(pet.getBreed())
                 .location(pet.getLocation())
+                .description(pet.getDescription())
+                .gender(pet.getGender().toString())
                 .imageUrls(imageUrls.stream().map(ImageUrl::getUrl).collect(Collectors.toList()))
                 .adopted(pet.isAdopted())
                 .owner(ownerEmail)
+                .ownerId(ownerId)
                 .build();
     }
 
@@ -110,7 +126,7 @@ public class PetService {
     public PageResponse<PetInfoPrivateResponse> getUserPets(String sortField, String sortDirection, int page, int size) {
         String username = SecurityContextHolder.getContext().getAuthentication().getName();
         User user = userRepository.findByEmail(username).orElseThrow(() -> new UserNotFoundException("User not found"));
-        String email = user.getEmail();
+
 
         List<String> allowedFields = List.of("name", "type", "breed", "location", "adopted");
         if (!allowedFields.contains(sortField)) {
@@ -132,7 +148,7 @@ public class PetService {
 
     @Transactional(readOnly = true)
     public PageResponse<PetInfoPublicResponse> getPetsWithPaginationAndSorting(int page, int size, String sortField, String sortDirection) {
-        List<String> allowedFields = List.of("name", "type", "breed", "location", "adopted");
+        List<String> allowedFields = List.of("name", "type", "breed", "location", "adopted", "gender");
         if (!allowedFields.contains(sortField)) {
             throw new IllegalArgumentException("Invalid sort field: " + sortField);
         }
@@ -154,74 +170,11 @@ public class PetService {
         return new PageResponse<>(petsPage);
     }
 
-    /**
-     * @deprecated This service will be deprecated soon as it does not match our api requirements
-     * Use the (pagination + sorted) services instead
-     */
-
-    public List<PetInfoPublicResponse> getAllPets() {
-        List<PetInfoPublicResponse> list = new ArrayList<>();
-        petRepository
-                .findByApproved(true)
-                .forEach(p -> {
-
-                    List<String> urls = new ArrayList<>(p.getImages().stream().map(ImageUrl::getUrl).toList());
-
-                    list.add(PetInfoPublicResponse.builder()
-                            .name(p.getName())
-                            .type(p.getType())
-                            .breed(p.getBreed())
-                            .location(p.getLocation())
-                            .imageUrls(urls)
-                            .adopted(p.isAdopted())
-                            .owner(p.getOwner().getEmail())
-                            .description(p.getDescription())
-
-                            .build());
-                });
-        return list;
-
-    }
-
-    /**
-     * @deprecated This service will be deprecated soon as it does not match our api requirements
-     * Use the (pagination + sorted) services instead
-     */
 
     @Transactional(readOnly = true)
-    public List<PetInfoPublicResponse> getPetsWithSorting(String field) {
-        List<String> allowedFields = List.of("name", "type", "breed", "location", "adopted");
-        if (!allowedFields.contains(field)) {
-            throw new IllegalArgumentException("Invalid sort field: " + field);
-        }
-
-        List<Pet> pets = petRepository
-                .findAll(((root, query, criteriaBuilder) -> criteriaBuilder.isTrue(root.get("approved"))), Sort.by(Sort.Direction.ASC, field));
-        List<PetInfoPublicResponse> list = new ArrayList<>();
-        for (Pet p : pets) {
-            List<String> imageUrls = p.getImages().stream().map(ImageUrl::getUrl).toList();
-
-            String ownerEmail = (p.getOwner() != null) ? p.getOwner().getEmail() : null;
-            list.add(PetInfoPublicResponse.builder()
-                    .name(p.getName())
-                    .type(p.getType())
-                    .breed(p.getBreed())
-                    .location(p.getLocation())
-                    .imageUrls(imageUrls)
-                    .adopted(p.isAdopted())
-                    .owner(ownerEmail)
-                    .description(p.getDescription())
-                    .build());
-        }
-        return list;
-
-
-    }
-
-
     public PageResponse<PetInfoPublicResponse> searchPets(String name, String type, String breed, String location, int page, int size, String sortField, String sortDirection) {
 
-        List<String> allowedFields = List.of("name", "type", "breed", "location", "adopted");
+        List<String> allowedFields = List.of("name", "type", "breed", "location", "adopted", "gender");
         if (!allowedFields.contains(sortField)) {
             throw new IllegalArgumentException("Invalid sort field: " + sortField);
         }
@@ -261,7 +214,7 @@ public class PetService {
 
     @Transactional(readOnly = true)
     public PageResponse<PetInfoPublicResponse> findPetsByExample(FindPetByExampleRequest request, int page, int size, String sortField, String sortDirection) {
-        List<String> allowedFields = List.of("name", "type", "breed", "location", "adopted");
+        List<String> allowedFields = List.of("name", "type", "breed", "location", "adopted", "gender");
         if (!allowedFields.contains(sortField)) {
             throw new IllegalArgumentException("Invalid sort field: " + sortField);
         }
@@ -279,6 +232,7 @@ public class PetService {
                 .type(request.getType())
                 .breed(request.getBreed())
                 .location(request.getLocation())
+                .gender(request.getGender() != null ? Gender.valueOf(request.getGender()) : null)
                 .adopted(request.getAdopted() != null ? request.getAdopted() : false)
                 .owner(request.getOwnerEmail() != null ?
                         User.builder().email(request.getOwnerEmail()).build() : null)
@@ -298,18 +252,6 @@ public class PetService {
         return new PageResponse<>(petsPage);
     }
 
-    /**
-     * @deprecated This service will be deprecated soon as it does not match our api requirements
-     * Use the (pagination + sorted) services instead
-     */
-    public PageResponse<PetInfoPublicResponse> getPetsWithPagination(int page, int size) {
-        Pageable pageable = PageRequest.of(page, size);
-        Page<PetInfoPublicResponse> petsPage = petRepository.findAll((root, query, cb) -> cb.isTrue(root.get("approved"))
-                        , pageable)
-                .map(PetInfoPublicResponse::fromEntity);
-        return new PageResponse<>(petsPage);
-
-    }
 
     //ALL WRITE SERVICES
 
@@ -318,8 +260,7 @@ public class PetService {
         String username = SecurityContextHolder.getContext().getAuthentication().getName();
         User user = userRepository.findByEmail(username)
                 .orElseThrow(() -> new UserNotFoundException("User not found"));
-        String email = user.getEmail();
-
+        Long userId = user.getId();
         Pet pet = petRepository.findById(petId)
                 .orElseThrow(() -> new PetListingNotFoundException("Listing not found"));
         List<Map<String, String>> imageUrls = pet.getImages().stream().map(img -> {
@@ -335,6 +276,15 @@ public class PetService {
         List<ImageUrl> imageUrl = pet.getImages();
         imageUrl.forEach(img -> cloudinaryService.deleteFile(img.getPublicId()));
         petRepository.delete(pet);
+
+        applicationEventPublisher.publishEvent(
+                NotificationEvent.builder()
+                        .recipientEmail(username)
+                        .recipientUserId(userId)
+                        .message("Your Pet listing " + pet.getName() + " was successfully removed and is no longer Live")
+                        .type(NotificationType.PET_DELETED).build()
+        );
+
         return PetInfoPrivateResponse.builder()
                 .id(pet.getId())
                 .name(pet.getName())
@@ -343,7 +293,7 @@ public class PetService {
                 .location(pet.getLocation())
                 .imageUrls(imageUrls)
                 .adopted(pet.isAdopted())
-                .owner(email)
+                .owner(username)
                 .build();
 
 
@@ -356,8 +306,9 @@ public class PetService {
     public PetInfoPrivateResponse addPetPost(AddPetRequest request) {
         String username = SecurityContextHolder.getContext().getAuthentication().getName();
         User user = userRepository.findByEmail(username).orElseThrow(() -> new UserNotFoundException("User not found"));
+        Long userId = user.getId();
         request.setOwner(user);
-        String email = user.getEmail();
+        Gender gender = Gender.valueOf(request.getGender());
 
 
         Pet savedPet = petRepository.save(Pet.builder()
@@ -365,9 +316,11 @@ public class PetService {
                 .type(request.getType())
                 .breed(request.getBreed())
                 .location(request.getLocation())
+                .gender(gender)
                 .description(request.getDescription())
                 .owner(request.getOwner())
                 .approved(true)
+                .approvedAt(LocalDateTime.now())
                 .build());
 
         List<ImageUrl> imageUrls = new ArrayList<>();
@@ -382,6 +335,14 @@ public class PetService {
         }
         petImageUrlsRepository.saveAll(imageUrls);
 
+        applicationEventPublisher.publishEvent(
+                NotificationEvent.builder()
+                        .recipientEmail(username)
+                        .recipientUserId(userId)
+                        .message("Your Pet listing " + savedPet.getName() + " is now Live")
+                        .type(NotificationType.PET_APPROVED).build()
+        );
+
         return PetInfoPrivateResponse.builder()
                 .id(savedPet.getId())
                 .name(savedPet.getName())
@@ -394,20 +355,61 @@ public class PetService {
                     return map;
                 }).toList())
                 .adopted(savedPet.isAdopted())
-                .owner(email)
+                .owner(username)
                 .description(savedPet.getDescription())
                 .build();
 
 
     }
 
+    @Transactional
+    public PetInfoPrivateResponse updatePetAdoptionStatus(long petId, Boolean adopted) {
+        String username = SecurityContextHolder.getContext().getAuthentication().getName();
+        User user = userRepository.findByEmail(username)
+                .orElseThrow(() -> new UserNotFoundException("User not found"));
+        Long userId = user.getId();
+
+
+        Pet pet = petRepository.findById(petId)
+                .orElseThrow(() -> new PetListingNotFoundException("Listing not found:" + petId));
+
+        if (!pet.getOwner().getEmail().equals(user.getEmail())) {
+            throw new UnauthorizedUserException("Pet Listing is not owned by User");
+        }
+        pet.setAdopted(adopted);
+        petRepository.save(pet);
+
+
+        applicationEventPublisher.publishEvent(
+                NotificationEvent.builder()
+                        .recipientUserId(userId)
+                        .recipientEmail(username)
+                        .message("Your Pet " + pet.getName() + " is now marked as adopted")
+                        .type(NotificationType.PET_UPDATED).build()
+        );
+
+
+        return PetInfoPrivateResponse.builder()
+                .id(pet.getId())
+                .name(pet.getName())
+                .type(pet.getType())
+                .breed(pet.getBreed())
+                .location(pet.getLocation())
+                .adopted(pet.isAdopted())
+                .owner(user.getEmail())
+                .description(pet.getDescription())
+                .build();
+
+
+    }
 
     @Transactional
     public PetInfoPrivateResponse updatePetPost(UpdatePetRequest request, long petId) {
         String username = SecurityContextHolder.getContext().getAuthentication().getName();
+
         User user = userRepository.findByEmail(username)
                 .orElseThrow(() -> new UserNotFoundException("User not found"));
-
+        Long userId = user.getId();
         Pet pet = petRepository.findById(petId)
                 .orElseThrow(() -> new PetListingNotFoundException("Listing not found:" + petId));
 
@@ -417,7 +419,7 @@ public class PetService {
 
         applyUpdates(pet, request);
 
-        if (request.getImageUrls() != null) {
+        if (request.getImageUrls() != null && !request.getImageUrls().isEmpty()) {
             // remove old images safely
             List<ImageUrl> oldImageUrls = pet.getImages();
             oldImageUrls.forEach(img -> cloudinaryService.deleteFile(img.getPublicId()));
@@ -437,6 +439,14 @@ public class PetService {
         }
 
         petRepository.save(pet);
+
+        applicationEventPublisher.publishEvent(
+                NotificationEvent.builder()
+                        .recipientUserId(userId)
+                        .recipientEmail(username)
+                        .message("Your Pet Listing " + pet.getName() + " was successfully updated ! Changes are now live.")
+                        .type(NotificationType.PET_UPDATED).build()
+        );
 
         return PetInfoPrivateResponse.builder()
                 .id(pet.getId())
@@ -460,6 +470,7 @@ public class PetService {
         if (request.getType() != null) pet.setType(request.getType());
         if (request.getBreed() != null) pet.setBreed(request.getBreed());
         if (request.getLocation() != null) pet.setLocation(request.getLocation());
+        if (request.getGender() != null) pet.setGender(Gender.valueOf(request.getGender()));
         if (request.getDescription() != null) pet.setDescription(request.getDescription());
         if (request.getAdopted() != null) pet.setAdopted(request.getAdopted());
     }
